+++ activerecord/CHANGELOG.md
@@ -1,3 +1,7 @@
+*   Remove deprecated `name` argument from `ActiveRecord::Base.remove_connection`.
+
+    *Rafael Mendonça França*
+
 *   Remove deprecated support to call `alias_attribute` with non-existent attribute names.
 
     *Rafael Mendonça França*